Implmented I repairable

+++ src/model/iRepairable.java
@@ -4,6 +4,6 @@ import java.util.ArrayList;
 
 public interface iRepairable
 {
-    public void setRepairables(ArrayList<Double> reparierte);
-    public ArrayList<Double> getRepairables();
+    public void setRepairables(ArrayList<Float> reparierte);
+    public ArrayList<Float> getRepairables();
 }
